--- pyproject.toml
@@ -1,26 +1,0 @@
-[project]
-name = "lingua"
-version = "0.1.0"
-description = "Add your description here"
-readme = "README.md"
-requires-python = ">=3.10"
-dependencies = [
-    "blobfile>=3.0.0",
-    "datatrove>=0.5.0",
-    "lm-eval>=0.4.9",
-    "msgspec>=0.19.0",
-    "ninja>=1.11.1.4",
-    "numpy>=2.2.6",
-    "omegaconf>=2.3.0",
-    "orjson>=3.10.18",
-    "pynvml>=12.0.0",
-    "rouge-score>=0.1.2",
-    "sacrebleu>=2.5.1",
-    "scipy>=1.15.3",
-    "sentencepiece>=0.2.0",
-    "tiktoken>=0.9.0",
-    "torch==2.7.0",
-    "viztracer>=1.0.4",
-    "wandb>=0.20.1",
-    #"xformers>=0.0.30",
-]
